# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sera Test API",
        "contact": {
            "name": "Ibnu Muntoha",
            "email": "ibnu.muntoha@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/restaurant/{id}": {
            "get": {
                "tags": [
                    "Restaurant Using MongoDB"
                ],
                "summary": "Get Restaurant By Id",
                "description": "Get Data Restaurant By spesicif document id",
                "operationId": "/retaurant/id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id parameter in path is document id in mongodb ex: SURvO6l5ILojZ9i3QeYfn8k0",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns restaurant data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "bool",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success Get Data"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "some restaurant  data"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Restaurant Using MongoDB"
                ],
                "summary": "Update Restaurant By Id",
                "description": "Update Restaurant data",
                "operationId": "Update Restaurant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Document id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "to update restaurant data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "borough": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "Jakarta"
                                    },
                                    "cuisine": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Betawi"
                                    },
                                    "restaurant_id": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "14025"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Solaria"
                                    },
                                    "address": {
                                        "properties": {
                                            "building": {
                                                "type": "string",
                                                "format": "text",
                                                "example": "230"
                                            },
                                            "street": {
                                                "type": "string",
                                                "format": "text",
                                                "example": "Jl.Kemerdekaan"
                                            },
                                            "zipcode": {
                                                "type": "string",
                                                "format": "text",
                                                "example": "14502"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "pet response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "bool",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success Updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Restaurant Using MongoDB"
                ],
                "summary": "Delete Restaurant By Id",
                "description": "Delete Data Restaurant By spesicif document id",
                "operationId": "/retaurant/id",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The id parameter in path is document id in mongodb ex: SURvO6l5ILojZ9i3QeYfn8k0",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns restaurant data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "bool",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success Deleted"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/restaurant/": {
            "post": {
                "tags": [
                    "Restaurant Using MongoDB"
                ],
                "summary": "Create Restaurant By Id",
                "description": "Creates a new Restaurant data",
                "operationId": "Add Restaurant",
                "requestBody": {
                    "description": "to add restaurant data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "borough": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "Jakarta"
                                    },
                                    "cuisine": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Betawi"
                                    },
                                    "restaurant_id": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "14025"
                                    },
                                    "name": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "Solaria"
                                    },
                                    "address": {
                                        "properties": {
                                            "building": {
                                                "type": "string",
                                                "format": "text",
                                                "example": "230"
                                            },
                                            "street": {
                                                "type": "string",
                                                "format": "text",
                                                "example": "Jl.Kemerdekaan"
                                            },
                                            "zipcode": {
                                                "type": "string",
                                                "format": "text",
                                                "example": "14502"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "pet response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "bool",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Success Created"
                                        },
                                        "_id": {
                                            "type": "string",
                                            "example": "azgZaZT7bCdInSAymsHgyVy0"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}